Update 2-uniq_add.py
This code defines a function uniq_add that takes a list of integers as input and returns the sum of all unique integers in the list. The function uses a set to keep track of the unique integers that have been seen so far. For each integer in the input list, the function checks if it is in the set. If it is not, the integer is added to the set and the result is incremented by the value of the integer.